move checkForUpdates
